Add visualization tests for mpc trajectory
