+++ CMakeLists.txt
@@ -413,6 +413,7 @@ catkin_add_gtest(${PROJECT_NAME}-uav-basic-state-machine-test tests/state_machin
 catkin_add_gtest(${PROJECT_NAME}-visual-servoing-state-machine-test tests/state_machines/visual_servoing_state_machine_tests.cpp)
 add_dependencies(${PROJECT_NAME}-visual-servoing-state-machine-test ${${PROJECT_NAME}_EXPORTED_TARGETS})
 add_rostest_gtest(${PROJECT_NAME}-system-handler-node-utils-test tests/common/system_handler_node_utils_tests.test tests/common/system_handler_node_utils_tests.cpp)
+add_rostest_gtest(${PROJECT_NAME}-mpc-trajectory-visualizer-test tests/common/mpc_trajectory_visualizer_tests.test tests/common/mpc_trajectory_visualizer_tests.cpp)
 add_rostest_gtest(${PROJECT_NAME}-uav-vision-system-test tests/robot_systems/uav_vision_system_tests.test tests/robot_systems/uav_vision_system_tests.cpp)
 add_rostest_gtest(${PROJECT_NAME}-state-machine-gui-connector-event-test tests/state_machines/state_machine_gui_connector_event_tests.test tests/state_machines/state_machine_gui_connector_event_tests.cpp)
 add_dependencies(${PROJECT_NAME}-state-machine-gui-connector-event-test ${${PROJECT_NAME}_EXPORTED_TARGETS} event_publish_node)
@@ -591,6 +592,9 @@ if(TARGET ${PROJECT_NAME}-roi-to-position-converter-test)
 endif()
 if(TARGET ${PROJECT_NAME}-system-handler-node-utils-test)
   target_link_libraries(${PROJECT_NAME}-system-handler-node-utils-test aerial_autonomy)
+endif()
+if(TARGET ${PROJECT_NAME}-mpc-trajectory-visualizer-test)
+  target_link_libraries(${PROJECT_NAME}-mpc-trajectory-visualizer-test aerial_autonomy)
 endif()
 if(TARGET ${PROJECT_NAME}-alvar-tracker-test)
   target_link_libraries(${PROJECT_NAME}-alvar-tracker-test aerial_autonomy)
